pyproject.toml
[tool.poetry]
name = "farpostbooks-telegram"
version = "0.1.0"
description = ""
authors = [
    "KatantDev <vlad26@mail.ru>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.0.0b7"
aiogram-dialog = "^2.0.0b16"
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
